duplicated script in a common dir, automatic mode names guess from file names, null protection
